Revert "* Resolved "signed/unsigned" Level 3 warnings."
This reverts commit 4681fb5718d571feeddb3de3ecd09911498db976.

+++ Trans3_2/vc/tkCommon/tkGfx/CUtil.h
@@ -69,7 +69,7 @@ inline std::string util::upperCase(std::string strString)
 {
 	std::string strToRet = "";
 
-	for (size_t c = 0; c < strString.length(); c++)
+	for (int c = 0; c < strString.length(); c++)
 	{
 		strToRet += toupper(strString.at(c));
 	}
